Remove MultiSignature (#127)
