+++ polkadot-js/alphanet-types.json
@@ -98,5 +98,7 @@
       "CallErrorAsFatal(ExitError)",
       "Other(Cow<'static, str>)"
     ]
-  }
+  },
+  "EthereumSignature": "[u8; 65]",
+  "ExtrinsicSignature": "EthereumSignature"
 }
