+++ runtime/src/parachain.rs
@@ -23,10 +23,10 @@ macro_rules! runtime_parachain {
 			spec_name: create_runtime_str!("moonbase-alphanet"),
 			impl_name: create_runtime_str!("moonbase-alphanet"),
 			authoring_version: 3,
-			spec_version: 5,
-			impl_version: 1,
+			spec_version: 6,
+			impl_version: 0,
 			apis: RUNTIME_API_VERSIONS,
-			transaction_version: 1,
+			transaction_version: 2,
 		};
 
 		impl cumulus_parachain_upgrade::Config for Runtime {
